requirements.txt
streamlit
SpeechRecognition
pygame
pillow
requests
